fix #3, #4  close #3, #4
Add new file "task_handler.py"

"task_handler.py" manages the saving, loading, and sorting of items in "tasks_archive.json".

If a user deletes a task element or list element (changes status from "In Work" to "Done") in the main frame (todo_app.py),
the new task_handler will scan the status of all elements in "tasks_archive.json" for "Done".
After the user adds a new element, the task_handler checks the tasks_archive for any main element with the status "Done" and cleans it automatically.

This will be further developed and optimized for another feature.